perf: unify benchmark fixtures
Move the transaction and mainnet replay benchmarks into a single evm2-eest bench binary and point CodSpeed at that one target. Convert the micro fixtures to the statetest transaction/post format and remove the revm comparison bench path.\n\nReplace the previous mainnet JSON zstd fixture, 5,078,680 bytes, with a freshly captured postcard binary zstd fixture for blocks 25347446-25347455, 4,255,712 bytes.
