+++ crates/eest/benches/evm/support.rs
@@ -1,0 +1,76 @@
+use crate::{cases::Bench, fixture::Suites};
+use criterion::{BatchSize, BenchmarkGroup, black_box, measurement::WallTime};
+use evm2::{
+    BaseEvmTypes, Evm, Precompiles, SpecId,
+    env::BlockEnv,
+    ethereum::{RecoveredTxEnvelope, ethereum_tx_registry},
+    evm::InMemoryDB,
+};
+
+type BenchEvm = Evm<BaseEvmTypes>;
+
+#[derive(Clone, Debug)]
+pub(crate) struct PreparedBench {
+    name: &'static str,
+    spec: SpecId,
+    block: BlockEnv,
+    db: InMemoryDB,
+    tx: RecoveredTxEnvelope,
+}
+
+impl PreparedBench {
+    pub(crate) fn load(bench: &Bench, suites: &Suites) -> Self {
+        let suite = suites.get(bench.fixture_path);
+        let case = suite.case(bench.name);
+        Self {
+            name: bench.name,
+            spec: bench.spec,
+            block: case.block(),
+            db: case.state(),
+            tx: case.tx(bench.spec),
+        }
+    }
+
+    pub(crate) fn sanity_check(&self) {
+        let mut runner = Runner::new(self);
+        let _ = runner.run().unwrap_or_else(|err| {
+            panic!("{} benchmark transaction must execute: {err:?}", self.name)
+        });
+    }
+
+    pub(crate) fn bench(&self, group: &mut BenchmarkGroup<'_, WallTime>) {
+        group.bench_function(format!("{}/transact", self.name), |b| {
+            b.iter_batched(
+                || Runner::new(self),
+                |mut runner| {
+                    black_box(runner.run().unwrap_or_else(|err| {
+                        panic!("{} benchmark transaction must execute: {err:?}", self.name)
+                    }))
+                },
+                BatchSize::SmallInput,
+            );
+        });
+    }
+}
+
+struct Runner {
+    evm: BenchEvm,
+    tx: RecoveredTxEnvelope,
+}
+
+impl Runner {
+    fn new(prepared: &PreparedBench) -> Self {
+        Self {
+            evm: new_evm(prepared.spec, prepared.block, prepared.db.clone()),
+            tx: prepared.tx.clone(),
+        }
+    }
+
+    fn run(&mut self) -> evm2::registry::HandlerResult<evm2::TxResult> {
+        self.evm.transact(&self.tx).map(evm2::ExecutedTx::commit)
+    }
+}
+
+fn new_evm(spec: SpecId, block: BlockEnv, db: InMemoryDB) -> BenchEvm {
+    Evm::new(spec, block, ethereum_tx_registry(spec), db, Precompiles::base(spec))
+}
