+++ crates/eest/src/fixture_io.rs
@@ -1,7 +1,7 @@
 use crate::{binary, blockchaintest::BlockchainTest};
 use std::{
     fs::{self, File},
-    io::{self, BufWriter, Read},
+    io::{self, BufWriter, Read, Write},
     path::Path,
 };
 use thiserror::Error;
@@ -39,7 +39,7 @@ pub enum FixtureWriteError {
 
 /// Returns true when the path uses the binary fixture extension.
 pub fn is_binary_path(path: &Path) -> bool {
-    path.extension().is_some_and(|extension| extension == "bin")
+    has_extension(path, "bin") || is_zstd_path(path) && stem_has_extension(path, "bin")
 }
 
 /// Reads a plain JSON fixture or zstd-compressed JSON fixture.
@@ -55,7 +55,7 @@ pub fn read_to_string(path: &Path) -> io::Result<String> {
 /// Reads a blockchain test fixture from JSON, zstd-compressed JSON, or postcard binary.
 pub fn read_blockchain(path: &Path) -> Result<BlockchainTest, FixtureReadError> {
     if is_binary_path(path) {
-        let bytes = fs::read(path)?;
+        let bytes = read_bytes(path)?;
         Ok(binary::from_bytes(&bytes)?)
     } else {
         let input = read_to_string(path)?;
@@ -66,7 +66,12 @@ pub fn read_blockchain(path: &Path) -> Result<BlockchainTest, FixtureReadError> 
 /// Writes a blockchain test fixture as JSON, zstd-compressed JSON, or postcard binary.
 pub fn write_blockchain(path: &Path, suite: &BlockchainTest) -> Result<(), FixtureWriteError> {
     if is_binary_path(path) {
-        fs::write(path, binary::to_vec(suite)?)?;
+        let bytes = binary::to_vec(suite)?;
+        if is_zstd_path(path) {
+            write_zstd_bytes(path, &bytes)?;
+        } else {
+            fs::write(path, bytes)?;
+        }
     } else {
         let writer = BufWriter::new(File::create(path)?);
         if is_zstd_path(path) {
@@ -80,6 +85,36 @@ pub fn write_blockchain(path: &Path, suite: &BlockchainTest) -> Result<(), Fixtu
     Ok(())
 }
 
+fn read_bytes(path: &Path) -> io::Result<Vec<u8>> {
+    if is_zstd_path(path) {
+        let file = File::open(path)?;
+        let mut decoder = ZstdDecoder::new(file)?;
+        let mut bytes = Vec::new();
+        decoder.read_to_end(&mut bytes)?;
+        Ok(bytes)
+    } else {
+        fs::read(path)
+    }
+}
+
+fn write_zstd_bytes(path: &Path, bytes: &[u8]) -> io::Result<()> {
+    let writer = BufWriter::new(File::create(path)?);
+    let mut encoder = ZstdEncoder::new(writer, ZSTD_COMPRESSION_LEVEL)?;
+    encoder.write_all(bytes)?;
+    encoder.finish()?;
+    Ok(())
+}
+
 fn is_zstd_path(path: &Path) -> bool {
-    path.extension().is_some_and(|extension| extension == "zst")
+    has_extension(path, "zst")
+}
+
+fn has_extension(path: &Path, extension: &str) -> bool {
+    path.extension().is_some_and(|candidate| candidate == extension)
+}
+
+fn stem_has_extension(path: &Path, extension: &str) -> bool {
+    path.file_stem().is_some_and(|stem| {
+        Path::new(stem).extension().is_some_and(|candidate| candidate == extension)
+    })
 }
