+++ crates/evm2/Cargo.toml
@@ -83,20 +83,12 @@ aws-lc-rs = { workspace = true, optional = true }
 
 [dev-dependencies]
 alloy-primitives = { workspace = true, features = ["serde", "std"] }
-criterion.workspace = true
 serde = { workspace = true, features = ["derive", "std"] }
 serde_json = { workspace = true, features = ["std"] }
 rand = { workspace = true, features = ["std"] }
 ark-std = { workspace = true }
 rstest.workspace = true
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
-
-[target.'cfg(not(target_family = "wasm"))'.dev-dependencies]
-revm = { workspace = true, features = ["test-types"] }
-
-[[bench]]
-name = "evm"
-harness = false
 
 [[example]]
 name = "custom_evm"
