+++ data/burntpix.json
@@ -53,16 +53,34 @@
         "storage": {}
       }
     },
-    "transaction": [
-      {
-        "data": "0xa4de9ab400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000064",
-        "gasLimit": "0x1dcd65000",
-        "gasPrice": "0x0a",
-        "nonce": "0x00",
-        "sender": "0x89d5e72a8a4a0330a65bbcef3032be2f728264a8",
-        "to": "0x49206861766520746f6f206d7563682074696d65",
-        "value": "0x00"
-      }
-    ]
+    "transaction": {
+      "data": [
+        "0xa4de9ab400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000064"
+      ],
+      "gasLimit": [
+        "0x1dcd65000"
+      ],
+      "gasPrice": "0x0a",
+      "nonce": "0x00",
+      "sender": "0x89d5e72a8a4a0330a65bbcef3032be2f728264a8",
+      "to": "0x49206861766520746f6f206d7563682074696d65",
+      "value": [
+        "0x00"
+      ]
+    },
+    "post": {
+      "Cancun": [
+        {
+          "expectException": null,
+          "indexes": {
+            "data": 0,
+            "gas": 0,
+            "value": 0
+          },
+          "hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+          "logs": "0x0000000000000000000000000000000000000000000000000000000000000000"
+        }
+      ]
+    }
   }
 }
