+++ data/curve-stableswap-2pool.json
@@ -61,17 +61,34 @@
         "storage": {}
       }
     },
-    "transaction": [
-      {
-        "data": "0x5b41b9080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000003635c9adc5dea000000000000000000000000000000000000000000000000000000000000000000000",
-        "gasLimit": "0xf4240",
-        "gasPrice": "0xa",
-        "nonce": "0x00",
-        "secretKey": "0x74aed4fbf80a899e5a050d36b05c3744b47c6a1dabc53418aba0188177490853",
-        "sender": "0x89d5e72a8a4a0330a65bbcef3032be2f728264a8",
-        "to": "0xcccccccccccccccccccccccccccccccccccccccc",
-        "value": "0x00"
-      }
-    ]
+    "transaction": {
+      "data": [
+        "0x5b41b9080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000003635c9adc5dea000000000000000000000000000000000000000000000000000000000000000000000"
+      ],
+      "gasLimit": [
+        "0xf4240"
+      ],
+      "gasPrice": "0xa",
+      "nonce": "0x00",
+      "sender": "0x89d5e72a8a4a0330a65bbcef3032be2f728264a8",
+      "to": "0xcccccccccccccccccccccccccccccccccccccccc",
+      "value": [
+        "0x00"
+      ]
+    },
+    "post": {
+      "Cancun": [
+        {
+          "expectException": null,
+          "indexes": {
+            "data": 0,
+            "gas": 0,
+            "value": 0
+          },
+          "hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+          "logs": "0x0000000000000000000000000000000000000000000000000000000000000000"
+        }
+      ]
+    }
   }
 }
