+++ data/erc20_transfer.json
@@ -21,16 +21,34 @@
         "storage": {}
       }
     },
-    "transaction": [
-      {
-        "data": "0x30627b7c",
-        "gasLimit": "0x7fffffffffffffff",
-        "gasPrice": "0x0",
-        "nonce": "0x0",
-        "sender": "0x1111111111111111111111111111111111111111",
-        "to": "0xcccccccccccccccccccccccccccccccccccccccc",
-        "value": "0x0"
-      }
-    ]
+    "transaction": {
+      "data": [
+        "0x30627b7c"
+      ],
+      "gasLimit": [
+        "0x7fffffffffffffff"
+      ],
+      "gasPrice": "0x0",
+      "nonce": "0x0",
+      "sender": "0x1111111111111111111111111111111111111111",
+      "to": "0xcccccccccccccccccccccccccccccccccccccccc",
+      "value": [
+        "0x0"
+      ]
+    },
+    "post": {
+      "Osaka": [
+        {
+          "expectException": null,
+          "indexes": {
+            "data": 0,
+            "gas": 0,
+            "value": 0
+          },
+          "hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+          "logs": "0x0000000000000000000000000000000000000000000000000000000000000000"
+        }
+      ]
+    }
   }
 }
