+++ data/onchain-lm-v2.json
@@ -252,16 +252,34 @@
         "storage": {}
       }
     },
-    "transaction": [
-      {
-        "data": "0x5a5c8c830000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000f00000000000000000000000000000000000000000000000000000000000000e6000000000000000000000000000000000000000000000000000000000000002a00000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000079a00000000000000000000000000000000000000000000000000000000000007f800000000000000000000000000000000000000000000000000000000000007a300000000000000000000000000000000000000000000000000000000000007ad0000000000000000000000000000000000000000000000000000000000000127000000000000000000000000000000000000000000000000000000000000014b00000000000000000000000000000000000000000000000000000000000007f9",
-        "gasLimit": "0x7fffffffffffffff",
-        "gasPrice": "0x0",
-        "nonce": "0x0",
-        "sender": "0x1111111111111111111111111111111111111111",
-        "to": "0x5f8e7d750e75b44747c058a204d8dea0d18fa5d3",
-        "value": "0x0"
-      }
-    ]
+    "transaction": {
+      "data": [
+        "0x5a5c8c830000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000f00000000000000000000000000000000000000000000000000000000000000e6000000000000000000000000000000000000000000000000000000000000002a00000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000079a00000000000000000000000000000000000000000000000000000000000007f800000000000000000000000000000000000000000000000000000000000007a300000000000000000000000000000000000000000000000000000000000007ad0000000000000000000000000000000000000000000000000000000000000127000000000000000000000000000000000000000000000000000000000000014b00000000000000000000000000000000000000000000000000000000000007f9"
+      ],
+      "gasLimit": [
+        "0x7fffffffffffffff"
+      ],
+      "gasPrice": "0x0",
+      "nonce": "0x0",
+      "sender": "0x1111111111111111111111111111111111111111",
+      "to": "0x5f8e7d750e75b44747c058a204d8dea0d18fa5d3",
+      "value": [
+        "0x0"
+      ]
+    },
+    "post": {
+      "Cancun": [
+        {
+          "expectException": null,
+          "indexes": {
+            "data": 0,
+            "gas": 0,
+            "value": 0
+          },
+          "hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+          "logs": "0x0000000000000000000000000000000000000000000000000000000000000000"
+        }
+      ]
+    }
   }
 }
